google mapを web nativeに分割
