+++ src/Place/Models/PDB01MPlace/Contexts/PCT011MPlacesList/ModelComponents/PMC01101GoogleMapPlacesList/GoogleMapPlacesList.tsx
@@ -1,13 +1,12 @@
-import { GoogleMap, Marker as MarkerWeb } from '@react-google-maps/api';
 import { useContext, useState } from 'react';
-import { Platform } from 'react-native';
 import MapView, { Marker, PROVIDER_GOOGLE } from 'react-native-maps';
 
-import { stylesWeb, stylesNative } from './GoogleMapPlacesListStyle';
+import { styles } from './GoogleMapPlacesListStyle';
 
 import { PCT011MPlacesList } from '@/Place/Models/PDB01MPlace/Contexts/PCT011MPlacesList';
 
 export const PMC01101GoogleMapPlacesList = () => {
+	console.log("native gogole mapが読み込み")
 	const { placesList } = useContext(PCT011MPlacesList);
 	const [ref, setRef] = useState<MapView | null>(null);
 
@@ -15,25 +14,9 @@ export const PMC01101GoogleMapPlacesList = () => {
 	const centerLat = center ? center.geometry.latitude : 35.6809591;
 	const centerLng = center ? center.geometry.longitude : 139.7673068;
 
-	if (Platform.OS === 'web') {
-		console.log('webが読み込まれた');
-
-		const centerPoint = {
-			lat: centerLat,
-			lng: centerLng,
-		};
-		return (
-			<GoogleMap mapContainerStyle={stylesWeb.containerStyle} center={centerPoint} zoom={10}>
-				{placesList.map((place) => {
-					const coordinate = { lat: place.geometry.latitude, lng: place.geometry.longitude };
-					return <MarkerWeb position={coordinate} title={place.name} key={place.place_id} />;
-				})}
-			</GoogleMap>
-		);
-	}
 	return (
 		<MapView
-			style={stylesNative.mapview}
+			style={styles.mapview}
 			provider={PROVIDER_GOOGLE}
 			ref={(r) => setRef(r)}
 			initialRegion={{
