+++ src/Place/Models/PDB01MPlace/Contexts/PCT011MPlacesList/ModelComponents/PMC01101GoogleMapPlacesList/GoogleMapPlacesListStyle.web.ts
@@ -1,8 +1,8 @@
 import { StyleSheet } from 'react-native';
 
 export const styles = StyleSheet.create({
-	mapview: {
+	containerStyle: {
+		height: '300px',
 		width: '100%',
-		height: 200,
 	},
 });
