+++ src/Place/Models/PDB01MPlace/Contexts/PCT011MPlacesList/ModelComponents/PMC01101GoogleMapPlacesList/index.ts
@@ -1,1 +1,2 @@
-export * from './GoogleMapPlacesList';
+// export * from './GoogleMapPlacesList';
+// export * from './GoogleMapPlacesList.web';
